refactor(@eurora/launcher): remove native command and instead use existing bits api
