+++ apps/docs/src/routes/components/launcher/+page.svelte
@@ -1,19 +1,20 @@
 <script lang="ts">
 	import { Calendar, Smile, Calculator, User, CreditCard, Settings, Plus } from '@lucide/svelte';
 
-	import { LauncherNative as Launcher } from '@eurora/launcher';
+	// import { LauncherNative as Launcher } from '@eurora/launcher';
+	import { Launcher } from '@eurora/launcher';
 	import { Badge } from '@eurora/ui';
 </script>
 
 <div>
 	<div class="launcher absolute left-1/2 top-1/4 w-[1100px] -translate-x-1/2">
 		<Launcher.Root class="rounded-lg border shadow-md">
-			<Launcher.Input placeholder="Search" class="ml-[180px]" />
-			<span class="absolute left-[0px] top-4 ml-2 mt-2 flex items-center gap-2">
+			<Launcher.Input placeholder="Search" />
+			<!-- <span class="absolute left-[0px] top-4 ml-2 mt-2 flex items-center gap-2">
 				<Badge class="h-[50px] w-fit"><Calendar /> Test 1</Badge>
 				<Badge class="h-[50px] w-fit"><CreditCard /> Test 2</Badge>
 				<Badge class="h-[50px] w-fit bg-black"><Plus /></Badge>
-			</span>
+			</span> -->
 			<Launcher.List>
 				<Launcher.Empty>No results found.</Launcher.Empty>
 				<Launcher.Group heading="Suggestions">
