+++ packages/custom-components/launcher/src/components/command/command-item.svelte
