+++ packages/custom-components/launcher/src/components/command/command-link-item.svelte
