+++ packages/custom-components/launcher/src/components/command/command-shortcut.svelte
