+++ packages/custom-components/launcher/src/components/input-area/index.ts
@@ -1,0 +1,2 @@
+import { default as Root } from './input-area.svelte';
+export { Root as InputArea };
