+++ packages/custom-components/launcher/src/index.ts
@@ -1,4 +1,3 @@
-import * as Launcher from './components/bits-command';
+import * as Launcher from './components/command';
 
-import * as LauncherNative from './components/native-command';
-export { Launcher, LauncherNative };
+export { Launcher };
