+++ packages/ui/src/index.ts
@@ -19,5 +19,3 @@ export * as Dialog from './components/ui/dialog/index.js';
 export * as Command from './components/ui/command/index.js';
 export { SystemChatMessage } from './custom-components/ui/system-chat-message/index.js';
 export { UserChatMessage } from './custom-components/ui/user-chat-message/index.js';
-
-export { AutoResizeTextarea } from './custom-components/ui/auto-resize-textarea/index.js';
